+++ src/DivisionalCouplerPanel.cpp
@@ -357,11 +357,17 @@ void DivisionalCouplerPanel::setDivisionalCoupler(DivisionalCoupler *divCplr) {
 		m_referencedSwitches->Enable(false);
 		m_addReferencedSwitch->Enable(false);
 		m_removeReferencedSwitch->Enable(false);
+		m_defaultToEngagedYes->Enable(true);
+		m_defaultToEngagedNo->Enable(true);
+		m_gcStateChoice->Enable(true);
 	} else {
 		m_availableSwitches->Enable(true);
 		m_referencedSwitches->Enable(true);
 		m_addReferencedSwitch->Enable(true);
 		m_removeReferencedSwitch->Enable(true);
+		m_defaultToEngagedYes->Enable(false);
+		m_defaultToEngagedNo->Enable(false);
+		m_gcStateChoice->Enable(false);
 	}
 	if (m_divCplr->isDefaultToEngaged())
 		m_defaultToEngagedYes->SetValue(true);
@@ -420,7 +426,7 @@ void DivisionalCouplerPanel::setTooltipsEnabled(bool isEnabled) {
 	if (isEnabled) {
 		m_displayInvertedYes->SetToolTip(wxT("Yes means that the on and off bitmaps are reversed."));
 		m_displayInvertedNo->SetToolTip(wxT("No is the standard on/off bitmap display."));
-		m_functionChoice->SetToolTip(wxT("Any other function than 'Input' means that the user no longer has direct control over the state of this element, instead it's controlled indirectly by referenced switch(es)."));
+		m_functionChoice->SetToolTip(wxT("Any other function than 'Input' means that the user no longer has direct control over the state of this element, nor can its state be directly changed by a combination, instead it's controlled indirectly by the referenced switch(es).\n\nThe logical function Not can only reference one switch and negates (inverts) the state of the input. And, Xor, Nand, Nor as well as Or has a variable number of inputs.\n\nSee for instance https://en.wikipedia.org/wiki/Logic_gate for further explanations about logical functions."));
 		m_defaultToEngagedYes->SetToolTip(wxT("The divisional coupler will be active by default. Only works if function is 'Input', otherwise the state is decided by switch(es)."));
 		m_defaultToEngagedNo->SetToolTip(wxT("The divisional coupler is not active by default."));
 		m_availableSwitches->SetToolTip(wxT("Switches available for referencing are listed here. If function is something else than 'Input' and there are switches available one or many can be selected for referencing."));
@@ -493,6 +499,9 @@ void DivisionalCouplerPanel::OnFunctionChange(wxCommandEvent& WXUNUSED(event)) {
 		m_referencedSwitches->Enable(false);
 		m_addReferencedSwitch->Enable(false);
 		m_removeReferencedSwitch->Enable(false);
+		m_defaultToEngagedYes->Enable(true);
+		m_defaultToEngagedNo->Enable(true);
+		m_gcStateChoice->Enable(true);
 	} else {
 		m_availableSwitches->Enable(true);
 		m_referencedSwitches->Enable(true);
@@ -502,6 +511,9 @@ void DivisionalCouplerPanel::OnFunctionChange(wxCommandEvent& WXUNUSED(event)) {
 				UpdateReferencedSwitches();
 			}
 		}
+		m_defaultToEngagedYes->Enable(false);
+		m_defaultToEngagedNo->Enable(false);
+		m_gcStateChoice->Enable(false);
 	}
 	::wxGetApp().m_frame->m_organ->setModified(true);
 }